refactor: spsp service to match service pattern
